requirements.txt
fastapi
uvicorn[standard]
cerebras_cloud_sdk
python-multipart
